+++ TDEngine2/include/platform/CProxyWindowSystem.h
@@ -84,6 +84,8 @@ namespace TDEngine2
 
 			TDE2_API E_RESULT_CODE SetScreenResolution(U32 width, U32 height) override;
 
+			TDE2_API E_RESULT_CODE SetIsFullscreenEnabled(bool state, bool borderlessMode = true) override;
+
 			/*!
 				\brief The method pass all the needed data into the immediate GUI context